+++ src/components/Routes.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Route, Switch } from 'react-router-dom'
 import { ProjectContainerConnected } from './project/ProjectContainer'
-import { ProjectsContainerConnected } from './Projects/Projects.container'
+import { ProjectsContainerConnected } from '../modules/Projects/Projects.container'
 import { ProjectCreateConnected } from './project/ProjectCreate'
 import { contentType } from '../types/models'
 import { ProjectForm } from './project/ProjectForm'
