+++ src/modules/Projects/ProjectCard.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { SDGArray, deviceWidth } from '../../lib/commonData'
 import styled from 'styled-components'
 import { Link } from 'react-router-dom'
-import { ProgressBar } from '../common/ProgressBar'
+import { ProgressBar } from '../../components/common/ProgressBar'
 import { excerptText } from '../../common/utils/formatters'
 
 const Title = styled.h3`
