+++ src/modules/Projects/Projects.container.tsx
@@ -2,16 +2,16 @@ import * as React from 'react'
 import styled from 'styled-components'
 import { ProjectCard } from './ProjectCard'
 import { ProjectsHero } from './ProjectsHero'
-import { Spinner } from '../common/Spinner'
+import { Spinner } from '../../components/common/Spinner'
 import { connect } from 'react-redux'
 import { RootState } from '../../common/redux/types'
 import { contentType } from '../../types/models'
-import { UserInfo } from '../../modules/account/types'
+import { UserInfo } from '../account/types'
 import { ProjectsDashboard } from './ProjectsDashboard'
-import FilterSortButtons from '../common/AdvancedFilter/FilterSortButtons'
+import FilterSortButtons from '../../components/common/AdvancedFilter/FilterSortButtons'
 import { deviceWidth } from '../../lib/commonData'
-import { fetchProjects } from '../../modules/Projects/Projects.actions'
-import { Project } from '../../modules/project/types'
+import { fetchProjects } from './Projects.actions'
+import { Project } from '../project/types'
 
 const Container = styled.div`
   display: flex;
