+++ src/modules/Projects/ProjectsDashboard.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import { WidgetWrapper, gridSizes } from '../common/WidgetWrapper'
-import { LayoutWrapper } from '../common/LayoutWrapper'
-import { SingleStatistic } from '../common/SingleStatistic'
+import { WidgetWrapper, gridSizes } from '../../components/common/WidgetWrapper'
+import { LayoutWrapper } from '../../components/common/LayoutWrapper'
+import { SingleStatistic } from '../../components/common/SingleStatistic'
 import { StatType } from '../../types/models'
 import { CircleProgressbar } from '../../common/components/Widgets/CircleProgressbar/CircleProgressbar'
 import { WorldMap } from '../../common/components/Widgets/WorldMap/WorldMap'
